refactor: `--ci` option
