+++ tests/.snapshots/success.txt
@@ -657,7 +657,7 @@
   ✓ it show the actual dataset of multiple non-named datasets in their description
   ✓ it show the correct description for mixed named and not-named datasets
 
-   PASS  Tests\Unit\Plugins\Context
+   PASS  Tests\Unit\Plugins\Environment
   ✓ environment is set to CI when --ci option is used
   ✓ environment is set to Local when --ci option is not used
